fix: Возврат /getAll без пагинации для совместимости с фронтендом
- /getAll возвращает List (как раньше) для фронтенда
- /getAllPaged добавлен с пагинацией для будущего
- /resetFilter возвращает List (как раньше)
- /count/{userId} исправлен на long
- JOIN FETCH оптимизация сохранена
- N+1 проблема устранена

+++ src/main/java/com/ulutman/controller/ManageCategoryController.java
@@ -79,17 +79,17 @@ public class ManageCategoryController {
         return manageCategoryService.getAllCategory();
     }
 
-
     @Operation(summary = "Count user publications")
     @ApiResponse(responseCode = "200", description = "Return count of the user's publications")
     @GetMapping("/count/{userId}")
-    public int getNumberOfPublications(@PathVariable Long userId) {
-        return (int) publishService.getNumberOfPublications(userId);
+    public long getNumberOfPublications(@PathVariable Long userId) {
+        return publishService.getNumberOfPublications(userId);
     }
-//    @Operation(summary = "Count user publications ")
-//    @ApiResponse(responseCode = "201", description = "Return count of the user's publications")
+
+//    @Operation(summary = "Count user publications")
+//    @ApiResponse(responseCode = "200", description = "Return count of the user's publications")
 //    @GetMapping("/count/{userId}")
 //    public int getNumberOfPublications(@PathVariable Long userId) {
-//        return publishService.getNumberOfPublications(userId);
+//        return (int) publishService.getNumberOfPublications(userId);
 //    }
 }
